+++ Koha/Plugin/Com/ByWaterSolutions/RapidoILL/scripts/sync_requests.pl
@@ -16,6 +16,7 @@
 # This program comes with ABSOLUTELY NO WARRANTY;
 
 use Modern::Perl;
+use utf8;
 
 use Getopt::Long;
 use Text::Table;