+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.1] - 2025-09-01
+
+### Fixed
+- [#88] Add utf8 pragma to sync_requests.pl for proper Unicode character display
+
 ## [0.9.0] - 2025-09-01
 
 ### Added